requirements.txt
Pillow==10.2.0
numpy=1.26.3
RPi.GPIO==0.7.1

inky==1.3.2

#spidev
#use "python3 -m pip install inky --no-dependencies" when not on a raspberry pi.